+++ tests/gcloud/test_GCloudStreaming.py
@@ -1,0 +1,6 @@
+import unittest
+
+
+# How to test this?
+class test_GCloudStreaming(unittest.TestCase):
+    pass