+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added AvroWriter class.
+- Added GCloudStreaming class.
+
+## [0.3.0] - xxxx-xx-xx
+
 
 ## [0.2.0] - 2019-11-05
 ### Added